🔨update Battle Process by UserDto

+++ src/main/java/com/battle/csbattle/service/SseService.java
@@ -2,6 +2,7 @@ package com.battle.csbattle.service;
 
 import com.battle.csbattle.battle.Battle;
 import com.battle.csbattle.battle.BattleType;
+import com.battle.csbattle.dto.UserDto;
 import com.battle.csbattle.util.SseUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
@@ -11,8 +12,8 @@ import java.util.concurrent.ConcurrentHashMap;
 
 @Service
 public class SseService {
-    private static final Long DEFAULT_TIMEOUT = 60L * 1000 * 60;                            // Sse 연결의 유효시간. 만료되면 자동으로 클라에서 재연결 요청. TODO: 게임에 걸리는 시간보다 길게 설정해두기
-    private static final Map<String, SseEmitter> waitingClients = new ConcurrentHashMap<>();       // 배틀 시작 전 대기 큐의 역할
+    private static final Long DEFAULT_TIMEOUT = 60L * 1000;                            // Sse 연결의 유효시간. 만료되면 자동으로 클라에서 재연결 요청. TODO: 게임에 걸리는 시간보다 길게 설정해두기
+    private static final Map<String, UserDto> waitingClients = new ConcurrentHashMap<>();       // 배틀 시작 전 대기 큐의 역할
     private final BattleService battleService;
 
     public SseService(BattleService battleService) {
@@ -23,7 +24,8 @@ public class SseService {
         //String id = userId + "_" + System.currentTimeMillis();                            // emitter 의 id 를 뜻함. 필요 시 사용
 
         SseEmitter emitter = new SseEmitter(0L);                               // emitter 생성 후 waiting clients 에 집어넣기
-        waitingClients.put(userId,emitter);
+        UserDto player=new UserDto(null,emitter,null);
+        waitingClients.put(userId, player);
 
         System.out.println("--- current clients : " + waitingClients);
         System.out.println("--- current clients size : " + waitingClients.size());
@@ -36,12 +38,17 @@ public class SseService {
             Battle onGoingBattle = battleService.findBattleOfUser(userId);
             if (onGoingBattle != null) {
                 for (String playerId : onGoingBattle.getPlayers().keySet()) {
-                    if (playerId != userId) {
-                        SseEmitter opponent = onGoingBattle.getPlayers().get(playerId);
+                    if (!playerId.equals(userId)) {
+                        SseEmitter opponent = onGoingBattle.getPlayers().get(playerId).getEmitter();
                         SseUtil.sendToClient(opponent, "opponent-left", userId + " 님이 게임을 나갔습니다.");
                     }
                 }
-                battleService.deleteBattleById(onGoingBattle.getId());
+                onGoingBattle.deletePlayerById(userId);
+                System.out.println(onGoingBattle.getPlayers());
+
+                if (onGoingBattle.getPlayers().size() == 0) {
+                    battleService.deleteBattleById(onGoingBattle.getId());
+                }
             }
         });
         emitter.onTimeout(() -> {
@@ -50,8 +57,8 @@ public class SseService {
         });
 
         for (String clientId : waitingClients.keySet()) {
-            SseEmitter checkingEmitter = waitingClients.get(clientId);
-            SseUtil.sendToClient(checkingEmitter,"checking-connection","checking connection");
+            SseEmitter checkingEmitter = waitingClients.get(clientId).getEmitter();
+            SseUtil.sendToClient(checkingEmitter, "checking-connection", "checking connection");
         }
 
         if (waitingClients.size() == 2) {                                                       // 배틀 생성 (배틀 유형은 한문제 풀기)
